Breaks infinite redirect loop when user logged in via social media or unrecognised email address

+++ app/controllers/auth_controller.rb
@@ -2,26 +2,13 @@ class AuthController < ApplicationController
   def login
     token = params[:token]
 
-    return redirect_to_api_gateway_login if token.nil?
+    redirect_to_api_gateway_login and return if token.blank?
 
     session[:user_token] = token
 
-    connect = Faraday.new(url: "#{ENV['API_URL']}") do |faraday|
-      faraday.request :url_encoded # form-encode POST params
-      faraday.response :logger # log requests to STDOUT
-      faraday.adapter Faraday.default_adapter # make requests with Net::HTTP
+    unless ensure_logged_in
+      redirect_to_api_gateway_login and return
     end
-
-    connect.authorization :Bearer, session[:user_token]
-    response = connect.get('/auth/check-logged');
-
-    if !response.status.to_s.match(/^20/)
-      session.delete(:user_token)
-      return redirect_to_api_gateway_login
-    end
-
-    session[:current_user] = JSON.parse response.body
-    session[:api_validation_ttl] = Time.now + Rails.configuration.session_revalidate_timer
 
     # TODO: Validate the user type
     redirect_url = session.delete(:return_to)
